april 14 ..

+++ src/test/java/Factile_Mode/DGBP_HSB_TestCase_23.java
@@ -40,11 +40,12 @@ public class DGBP_HSB_TestCase_23 extends Base {
 	private static String filePath = System.getProperty("user.dir") + "\\src\\main\\java\\images\\eagle.jpg";
 
 	@BeforeTest
-	public void initilize() throws IOException {
+	public void initilize() throws IOException, InterruptedException {
 		driver = IntilizeDriver();
 		 Dimension d = new Dimension(1382,744);
 		 driver.manage().window().setSize(d);
 		Log.info("Driver is Initilize");
+		Thread.sleep(3000);
 		driver.get(prop.getProperty("rooturl"));
 		wait = new WebDriverWait(driver, 60);
 		Log.info("Navigated to homePage");
@@ -81,7 +82,16 @@ public class DGBP_HSB_TestCase_23 extends Base {
 		 driver1.quit();
 		 driver.switchTo().window(driver.getWindowHandle());
 		 driver.quit();
-		 Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		 String osName = System.getProperty("os.name");
+		 if (osName.equals("Windows 10")) {
+		 		Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		 		}
+		 		else
+		 		{
+		 			String[] cmd = new String[]{"/bin/sh", "killchrome.sh"};
+		 			Process pr = Runtime.getRuntime().exec(cmd);
+		 		}
+
 	}
 
 	public void modeaterscreen() throws InterruptedException {
